requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pymupdf==1.24.5
langchain==0.2.6
langchain-community==0.2.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
groq==0.9.0
streamlit==1.36.0
requests==2.32.3
ragas==0.1.14
datasets==2.20.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1